Add Gaussian noise function
This function generates samples from a gaussian distribution
given the mean and standard deviation.
It will be used to generate the points for the monte carlo
simualtion's random price path.
This number assumes that each price point is independant of
the last (not true in reality) since it is drawing from a normal
distribution.