Add the monte carlo simluation
For every iteration of the simulation calculate the price path.
The price path is the simulated value of the option at the time
of maturity(assumed to be 1 year).
The calculated price is the checked for whether it is a call
option or not
the payoff is then calculated and avergaed out for all the simulations
The rate and sigma hyperparameters should usually be dynamic,
changing for each iteration. The distribution of the sample
chosen for calcultion can also be changed from gaussian to run
different simultions and figure out the payout.
[Ticket: X]